requirements.txt
Python==3.13.1
Flask==3.1.0
psycopg2-binary==2.9.10